+++ backend/src/models/change_request.js
@@ -43,7 +43,8 @@ const findById = (changeRequestId, projectId) => {
     .from(changeRequestTable)
     .leftJoin(fiscalYearTable, { "change_request.fiscal_year": `${fiscalYearTable}.id` })
     .where({ "change_request.id": changeRequestId })
-    .where({ "change_request.link_id": projectId });
+    .where({ "change_request.link_id": projectId })
+    .first();
 };
 
 // Update one.