+++ frontend/src/hooks/useFormatTableData.tsx
@@ -56,7 +56,7 @@ export const formatTableColumns = (
         field: "edit",
         headerName: "",
         sortable: false,
-
+        filterable: false,
         maxWidth: 60,
         renderCell: (cellValues: { id: number }) => {
           return (
@@ -91,6 +91,8 @@ export const formatTableColumns = (
           .replace(/(?:^|\s)\S/g, (a: string) => a.toUpperCase()),
         flex: getFlexValue(value[0]),
         id: index,
+        sortable: false,
+        filterable: false,
       });
     });